+++ packages/mpack/src/bundler/plugins/transformPlugin/steps/createFullyTransformStep.ts
@@ -45,13 +45,13 @@ export function createFullyTransformStep({
        * @babel/preset-react 포함 필요
        */
       [require.resolve('@babel/preset-react'), { runtime: 'automatic' }],
+      [require.resolve('@react-native/babel-preset')],
       ...(additionalBabelOptions?.presets ?? []),
     ],
     plugins: [
       /**
        * react-native에서 직접 export 하는 flow 파일 대응을 위해 strip types 추가 필요
        */
-      require.resolve('@babel/plugin-transform-flow-strip-types'),
       [require.resolve('@babel/plugin-proposal-class-properties'), { loose: true }],
       [require.resolve('@babel/plugin-proposal-private-property-in-object'), { loose: true }],
       [require.resolve('@babel/plugin-proposal-private-methods'), { loose: true }],