+++ services/showcase/granite.config.ts
@@ -2,7 +2,6 @@ import { env } from '@granite-js/plugin-env';
 import { hermes } from '@granite-js/plugin-hermes';
 import { microFrontend } from '@granite-js/plugin-micro-frontend';
 import { router } from '@granite-js/plugin-router';
-import { sentry } from '@granite-js/plugin-sentry';
 import { defineConfig } from '@granite-js/react-native/config';
 
 export default defineConfig({
@@ -15,7 +14,6 @@ export default defineConfig({
     router(),
     env({ MY_ENV: 'from granite.config.ts' }),
     hermes(),
-    sentry({ useClient: false }),
     microFrontend({
       name: 'remoteApp',
       exposes: {